amivbooklet: Switch to multicol
The twocolumn environment available by latex is unfortunately not well suited
to put full page images in.
The multicol package allows switching column mode everywhere on the page,
therefore allowing to deactivate the twocolumn mode to place a full width
picture and activating it again.

To make interaction with it as easy as possible, automatic activation
of the environment has been added (still experimental)